# frontend/src/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: TinBoker Backend API
  description: Backend API for TinBoker - Financial podcast insights platform
  version: 1.0.0
servers:
- url: https://api.tinboker.com
  description: Production server
- url: https://staging-api.tinboker.com
  description: Staging server
- url: https://dev-api.tinboker.com
  description: Development server
- url: http://localhost:8000
  description: Local development
paths:
...
  /api/podcast/{podcast_name}/episodes/{episode_id}:
    get:
      tags:
      - podcast
      summary: Get Episode By Id
      description: 'Get specific episode by ID


        Returns complete episode data including transcript, summary, and related tickers


        CDN Cache: 30 minutes'
      operationId: get_episode_by_id_api_podcast__podcast_name__episodes__episode_id__get
      parameters:
      - name: podcast_name
        in: path
        required: true
        schema:
          type: string
          description: Podcast name
          title: Podcast Name
        description: Podcast name
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          description: Episode ID
          title: Episode Id
        description: Episode ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Episode'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/episodes/recent:
    get:
      tags:
      - episodes
      summary: Get Recent Episodes
      description: 'Get recent episodes across all podcasts, sorted by created_time
        descending


        Query params:

        - limit: Maximum number of episodes to return (default: 20, max: 200)

        - offset: Pagination offset

        - podcast_name: Optional filter by podcast name


        CDN Cache: 30 minutes'
      operationId: get_recent_episodes_api_episodes_recent_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Maximum number of episodes to return (1-200)
          default: 20
          title: Limit
        description: Maximum number of episodes to return (1-200)
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Pagination offset
          default: 0
          title: Offset
        description: Pagination offset
      - name: podcast_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional filter by podcast name
          title: Podcast Name
        description: Optional filter by podcast name
      - name: include_content
        in: query
        required: false
        schema:
          type: boolean
          description: Include heavy content fields (transcript, summary)
          default: false
          title: Include Content
        description: Include heavy content fields (transcript, summary)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/episodes/by-ticker/{ticker}:
    get:
      tags:
      - episodes
      summary: Get Episodes By Ticker
      description: 'Get episodes that mention a specific stock ticker


        Path params:

        - ticker: Stock ticker symbol (e.g., "NVDA", "AAPL")


        Query params:

        - limit: Maximum number of episodes to return (default: 50, max: 200)

        - offset: Pagination offset


        CDN Cache: 30 minutes'
      operationId: get_episodes_by_ticker_api_episodes_by_ticker__ticker__get
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          description: Stock ticker symbol
          title: Ticker
        description: Stock ticker symbol
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Maximum number of episodes to return (1-200)
          default: 50
          title: Limit
        description: Maximum number of episodes to return (1-200)
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Pagination offset
          default: 0
          title: Offset
        description: Pagination offset
      - name: include_content
        in: query
        required: false
        schema:
          type: boolean
          description: Include heavy content fields (transcript, summary)
          default: false
          title: Include Content
        description: Include heavy content fields (transcript, summary)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Episode:
      properties:
        id:
          type: string
          title: Id
          description: Unique episode identifier
        podcast_name:
          type: string
          title: Podcast Name
          description: Podcast name
        episode_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Episode Title
          description: Episode title
        episode_number:
          anyOf:
          - type: integer
          - type: 'null'
          title: Episode Number
          description: Episode number
        transcript:
          anyOf:
          - type: string
          - type: 'null'
          title: Transcript
          description: Episode transcript
        summary_content:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary Content
          description: Summary markdown content
        summary_image:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary Image
          description: Summary SVG image
        related_tickers:
          items:
            type: string
          type: array
          title: Related Tickers
          description: Related stock tickers
        tags:
          items:
            type: string
          type: array
          title: Tags
          description: Topic tags
        created_time:
          type: integer
          title: Created Time
          description: Creation timestamp (Unix milliseconds)
        number_click:
          type: integer
          title: Number Click
          description: Number of clicks
          default: 0
        num_likes:
          type: integer
          title: Num Likes
          description: Number of likes
          default: 0
        key_insights:
          items:
            type: string
          type: array
          title: Key Insights
          description: Key insights for the episode
        raw_mp3:
          anyOf:
          - type: string
          - type: 'null'
          title: Raw Mp3
          description: Raw MP3 file path (local, not in Firebase)
        mp3_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Mp3 Url
          description: GCS URL for MP3 file (gs://...)
        transcript_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Transcript Url
          description: GCS URL for transcript file (gs://...)
        summary_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary Url
          description: GCS URL for summary markdown file (gs://...)
        summary_image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary Image Url
          description: GCS URL for SVG image file (gs://...)
        events_markdown_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Events Markdown Url
          description: GCS URL for events markdown file (gs://...)
        sentences_markdown_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Sentences Markdown Url
          description: GCS URL for sentences markdown file (gs://...)
        marp_markdown_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Marp Markdown Url
          description: GCS URL for Marp markdown file (gs://...)
        mp3_public_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Mp3 Public Url
          description: Public HTTPS URL for MP3
        transcript_public_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Transcript Public Url
          description: Public HTTPS URL for transcript
        summary_public_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary Public Url
          description: Public HTTPS URL for summary
        summary_image_public_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary Image Public Url
          description: Public HTTPS URL for SVG image
        events_markdown_public_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Events Markdown Public Url
          description: Public HTTPS URL for events markdown
        sentences_markdown_public_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Sentences Markdown Public Url
          description: Public HTTPS URL for sentences markdown
        marp_markdown_public_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Marp Markdown Public Url
          description: Public HTTPS URL for Marp markdown
        events_markdown_content:
          anyOf:
          - type: string
          - type: 'null'
          title: Events Markdown Content
          description: Events markdown content
        sentences_markdown_content:
          anyOf:
          - type: string
          - type: 'null'
          title: Sentences Markdown Content
          description: Sentences markdown content
        marp_markdown_content:
          anyOf:
          - type: string
          - type: 'null'
          title: Marp Markdown Content
          description: Marp markdown content
        ticker_marp_markdown_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Ticker Marp Markdown Url
          description: GCS URL for ticker-specific Marp markdown file (gs://...)
        ticker_marp_markdown_public_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Ticker Marp Markdown Public Url
          description: Public HTTPS URL for ticker-specific Marp markdown
        ticker_marp_markdown_content:
          anyOf:
          - type: string
          - type: 'null'
          title: Ticker Marp Markdown Content
          description: Ticker-specific Marp markdown content
        ticker_insights_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Ticker Insights Url
          description: GCS URL for ticker insights JSON file (gs://...)
        ticker_insights_public_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Ticker Insights Public Url
          description: Public HTTPS URL for ticker insights JSON
        ticker_insights_content:
          anyOf:
          - type: string
          - type: 'null'
          title: Ticker Insights Content
          description: Ticker insights JSON content (cached)
        spotify_embed_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Spotify Embed Url
          description: Spotify embed URL
        spotify_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Spotify Id
          description: Spotify episode ID
        spotify_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Spotify Url
          description: Spotify episode URL
        spotify_release_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Spotify Release Date
          description: Release date from Spotify (YYYY-MM-DD)
        spotify_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Spotify Description
          description: Episode description from Spotify
        spotify_duration_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Spotify Duration Ms
          description: Episode duration in milliseconds
        spotify_images:
          items:
            type: string
          type: array
          title: Spotify Images
          description: List of image URLs from Spotify
        modified_summary_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Modified Summary Url
          description: GCS URL for modified summary markdown file (gs://...)
        modified_summary_content:
          anyOf:
          - type: string
          - type: 'null'
          title: Modified Summary Content
          description: Modified summary markdown content
        modified_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Modified By
          description: User who modified the summary (email or ID)
        modified_at:
          anyOf:
          - type: integer
          - type: 'null'
          title: Modified At
          description: Timestamp when summary was modified (Unix milliseconds)
      type: object
      required:
      - id
      - podcast_name
      - created_time
      title: Episode
      description: Podcast episode model
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer